Add ClientLease implementation and tests.
ClientLease is used by the client library to abstract the process
of aquiring and maintaining a valid lease with the coordinator.
These leases are used to represent the lifetime of an active client
so that per client state stored on servers (e.g. linearizability
data) can be garbage collected when clients fail or become inactive.