End-to-end implementation of split and migrate indexlet.
Client API (in RamCloud class) allows caller to provide an indexlet and
the key for the split (by table Id, index Id, split key, split key length),
as well as the Id of the server where the second indexlet resulting from
this split should be migrated. The first indexlet remains on the original
owner of indexlet that was split.

Also made the following changes:
1. Remove lint where backing tables (and tablets) for indexlets were referred
to as indexTable or indexletTable (/ indexTablet or indexletTablet) and
renamed to backingTable (/ backingTablet).
2. Use similar nomenclature to get/find indexlet based on the index key range
or containing a given index key in IndexletManager and TableManager classes.
3. Change Indexlet structure to have the keys as void* rather than const void*
4. In TableManager::createTable(), allow the caller to specify the master
server on which to locate tablets for the table being created.
5. Add state field to IndexletManager::Indexlet that indicates whether a
given indexlet is NORMAL or RECOVERING as well as corresponding functions.
The RECOVERING state is used when an indexlet is either recovering or
being migrated from another server to this server. Modify recovery code to use
this state.
6. Change highestUsedId to nextNodeId in IndexletManager code for simplicity.
(nextNodeId is generally highestUsedId + 1.)
7. Reorder IndexletManager functions to follow convention.
8. Improve some documentation, add some missing unit tests, improve some
existing unit tests for SLIK code.
9. Improve code (documentation, naming, order and other things) for issues
discovered in code not related to SLIK.
10. Cleaner / more efficient implementation for IndexletManager::getIndexlet()
with better semantics for when InternalError is thrown.

Still to be done:
1. End-to-end system tests.
2. Benchmarking.
3. Coordinator fault-tolerance for this function.