Add TxRecoveryManager daft module with tests (incomplete).
The TxRecoveryManager is a module that runs on the master that is
charged with recovering transactions that have only partially
completed.  This commit includes the code that manages recovery
but not the code to recover from a recovery failure.  This commit
also includes unit tests for part of the recovery management but
leaves some to be added later once the corresponding transaction
code has been add to handle the requests generated by this manager.