Fix bug caused by some corner cases in IndexLookup.
In IndexLookup class, handle case where number of objects returned
for a primary key hash is greater than one as well as the case
where no objects are returned for a primary key hash.
Most of the infrastructure to handle both of these cases was
present previously, but there was a subtle bug which this
commit tries to fix.