Allow transactions to lock non-preexisting objects
Reimplement the LockTable and its callers to use PreparedOp objects
instead of Object objects to represent a unique lock.  This allows
transactions to lock objects that do not exist before the transaction.

The LockTable is optimized so that the following common cases are fast
assuming the table is sparsely populated:
  - Checking if lock is acquired when it is not.
  - Acquiring a lock when it is not already acquired.
  - Releasing a lock.
In these common cases, no full key comparisons should be need.  Full
key comparisons are needed, for example, when checking if lock is
acquired when it is indeed already locked (when a write request comes
in during an ongoing transaction).  Slowing this write path down is
acceptable since these operations will need to be retried anyway.

RAM-710 #resolve