Fix possible ClientTransactionTask "double free"
Construct ClientTransactionTask::Poller object using manual placement
new instead of a Tub so that the Poller's destructor is not invoked
when the ClientTransactionTask's destructor is called.  In the case
where a Transaction object is destructed before two phase commit is
completed by the ClientTransactionTask, the Poller's later destruction
will trigger the destruction of the ClientTransactionTask. If the
Poller is managed as a Tub, destructing the ClientTransactionTask will
cause the Poller to be destructed again (double free). The double free
is now avoided since the ClientTransactionTask's destructor does not
trigger the Poller's destructor.