Add support for Transaction::ReadOp batching.
A new bool is added to the Transaction::ReadOp constructor that allows
the client to indicate whether the ReadOp can be batched. Any ReadOps
constructed with 'batch == true' will not be sent immediately. Instead,
it will be collected into a batched MultiRead until ReadOp::isReady()
is called; isReady will cause the current batch to be flushed.