Schedule worker threads based on RPC levels, not services.
This changes the basic worker thread scheduling mechanism to eliminate
the potential for distributed deadlock, which was there previously.
The new system does not support per-service limits on concurrency;
it only supports a single overall thread limit.  The --masterServiceThreads
command-line option has been replaced with --maxCores (which includes
the dispatch thread as well as worker threads). Similarly, for the
coordinator, the old --threads option has been replaced with --maxCores.
Note: this means that all services are now concurrent, whereas in the
past some services had a concurrency limit of 1; this could expose latent
threading issues that were hidden previously by the lack of concurrency.