Remove Dispatch Poller use from Transaction commit
It is not safe for the transaction commit mechanism to be driven from
the poller as it may causes deadlock; RAM-770 shows a similar problem.
ClientTransactionTask now "steals cycles" from any calls to
Transaction::commit (i.e. calling commit on transaction B will also
drive the async commit of a preceding transaction A).  The module
ClientTransactionManager is reintroduced to track incomplete async
transaction commits.