Add tablet key range tracking to MasterService
MasterService is instrumented with TableStats calls to track owned
table key hash counts.  Every call to TabletManager::addTablet or
TabletManager::deleteTablet should be accompanied with a call to
TableStats::addKeyHashRange or TableStats::deleteKeyHashRange
respectively.  TabletManager::deleteTablet is also modified to return
whether or not a tablet was actually deleted (similar to addTablet);
this is useful for determining whether or not deleteKeyHashRange
should be called.