Reworked log iteration to fix migration bugs.
* Log iteration no longer grabs the append lock (this could cause deadlocks
and it isn't needed); major refactoring of LogIterator::next.
* During migration, don't check hash table to see if objects are still
live; this can cause an object to be missed if it has been moved to a
side segment by the cleaner, and the side segment hasn't yet been added
back to the log.
* Eliminated LogIterator::getPosition, LogIterator::setPosition,
LogIterator::refresh (no longer needed).
* Added new method SegmentIterator::setLimit.
* Modified epoch mechanism to support different classes of activity (migration
needs to wait for appends to complete, but meeting to wait for read-only
accesses, such as other migrations).
* Eliminated Transport::INVALID_EPOCH, switched to use 0 as an invalid
epoch.

RAM-783 #close RAM-784 #close