Merge stats from SideLogs into Log on commit.
SideLogs weren't updating main log stats when they committed. This was
leading to a case in migration that was causing underflow on
totalLiveBytes when tables were dropped after migration, which would
cause the log to believe it was massive and would freeze all writes.

SideLogs now merge their stats into the main log when they commit.

Fixes RAM-787.