Fixed Object::appendValueToBuffer()'s API
Formerly, Object::appendValueToBuffer() had the optional parameter valueOffset,
but in all uses of the function, the value passed in was the result of
Object::getValueOffset() and not specifying the parameter would result in
incorrect behavior. Thus to prevent future misuse, the parameter has been
removed and getValueOffset() call is integrated within appendValueToBuffer().