Ordered TabletManager::toString's output for tests
MasterServiceTest, ObjectManagerTest, and TabletManagerTest do string
comparison tests withmTabletManager::toString() which uses an iterator over a
std::unordered_multimap TabletMap to produce its output. The problem here is
since the data structure is unordered, the order of traversal with an iterator
is not guaranteed. This means the tests can fail due to a differing order, even
if the test should pass semantically. To fix this issue, TabletManager's
toString now returns the tablets in sorted order during debug=yes.