Changed PerfCounterTest's usleep to Cycles::sleep
Several tests in PerfCounterTest used usleep() when performing timing tests.
This is problematic as it may signal to the operating system that it's okay
to deschedule the thread and cause the timings to be off. To lower the
chance this happening, all the usleeps have been changed to Cycles::sleep
in PerfCounterTest.cc