Reset Cycles::cyclesPerSec in ClientSessionTests
ClientSessionTest::init sets Cycles::cyclesPerSec to a different value for
testing but does not reset it to the default value after use. This causes
subsequent tests that depend on the Cycles::toSeconds() to fail as it will
return the wrong value. This patch fixes the tests so they restore the value.