Fixed the remaining test failures on gcc 4.9.2
All the unit tests that were failing on the test Debian machine with gcc 4.9.2
should now pass. The two fixes specific to this commit are:

1) RecoveryTest.partitionTablets_mixed and TableStatsTest.estimator_constructor
had EXPECT's that relied on the output order of std::unordered_map::iterator,
which is not garunteed to return objects in any particular order. The trivial
fix was to sort the items tested prior to any EXPECT/ASSRT's.

2) SegmentIteratorTest.next was /incorrectly/ passing on gcc 4.4.7 due to
memory regions not being cleared before use. The two part fix was to first make
the test fail consistenty by zeroing out the segletBlocks and then fixing the
tests so they pass. The issue of not clearing the memory before use has been
filed as a jira bug ( RAM-794 ).