Fixes skipped nodes while migrating index.
Once the migration reaches the head segment, it truncates the
indexlet being migrated so this server no longer serves keys
starting with split key. After this, migration processes new log
entries in the head segment and those added in the meantime.
As part of this processing the older code required looking up
an indexlet using the split key, which fails as the indexlet
was just trucated.

Now, this log entry processing uses the information about
RAMCloud objects encapsulating the index tree node.
This avoids the above problem as well as increases efficiency by
reducing the amount of redundant work done.