GCC 4.9.2 compatible; added CONST_DOUBLE macro
RAMCloud now compiles on gcc 4.9.2 without modifications and tests should
pass with the same frequency as they did on 4.4.7. The last modification
was providing a CONST_DOUBLE macro which resolves into a "constexpr double"
when the compiler supports (and expects) it and "const double" when the
compiler does not. This roughly translates to a cutoff point of gcc 4.6
and above for constexpr.