Add ClusterTime class for clarity.
This is the first step to removing the uint64_t representation of
cluster-time.  This commit defines a ClusterTime class to represent
cluster-time and provides defined operations that can be performed
on cluster-time.  The goal is to prevent accidental miss-use of
cluster-time that could occur given its current unprotected uint64_t
representation.  This commit only adds the class; it does not add
use of the class.