Optimize UnackedRpcResults lease validation
UnackedRpcResults::checkDuplicate now uses the most up-to-date lease
information available to it when validating the lease (it previously
used the provided lease parameter which may not be the most current).
This optimization avoids the need to perform the more expensive lease
validation step involving a coordinator RPC even when the caller does
not have the most up-to-date lease.