More robust unit testing and bug fix in splitAndMigrateIndexlet.
This commit makes the end-to-end unit test for splitAndMigrateIndexlet in
TableManagerTest more robust. In addition to checking the correctness of
metadata, it now also migrates real index entries and tests that they are
migrated correctly.

The additional testing uncovered a bug whose symptom was that the index
entries that should have been migrated to a given server were not visible
on that server. This was because the status of the new backing table was set
to RECOVERING before the migration began but never set back to NORMAL after
the migration completed. This commit fixes this bug.

Finally, comments are added to the unit test (mentioned above) to make it
more readable.