RAM-808 #close (git pre-commit hooks failing)
Before this commit, hooks/pre-commit-20test would fail when run as
a part of a "git commit" with gradlew complaining that g++ was unable
to find "SpinLockStatistics.pb.h". The core issue (documented on jira)
turned out to be git setting the environment variable GIT_DIR when
executing hooks and the git commands in gradlew using it verbatim,
even after changing directories. This commit resolves this issue by
having the Makefile invoke gradlew without changing directories.