Fix PreparedItem::handleTimerEvent deadlock
PreparedItem::handleTimerEvent() previously synchronously sent the
TxHintFailed RPC and held a thread and worker timer while waiting.
This can causes a deadlock when the RPC is destined to loop back to
the same master (e.g. all threads are busy waiting for the worker
timer to be released and there are no threads available to process the
TxHintFailed RPC).  The solution is to make the handler asynchronous
using a rules-like structure; the handler will asynchronously send the
RPC and rescheduled itself polling for the RPC's completion.