Clean-up ClientTransactionTask with C++11 features
The full C++11 spec, which we now support, allows for the emplace
operation on std::unordered_map.  This allows the commit cache
CacheEntry to be emplaced in the commit cache map instead of being
copied.  It also allows the CacheEntry's object buffer to exist in
the entry as a member as opposed to a pointer to a dynamically
allocated and manually memory managed buffer.  These two improvements
serve to simplify the code.

RAM-733 #close