Set up Travis Continuous Integration
Travis CI is a free cloud service that will pull commits from Github,
perform user-specified tests, and e-mail the results. Currently, it's
set up so that on a push to master, it will make RAMCloud (w and w/o
DEBUG) and make tests/check/docs. The tests run on Ubuntu 12.04 VMs,
so it should at least help us check compatibility on platforms other
than Debian.

The only limitation is that the tests run on whimpy "1.5" core VMs.
This means that it takes about 15 minutes to compete all the tests and
two timing dependent tests had to be omitted (PortAlarmTest.triple_alarm
and WorkerTimerTest.stopInternal_handlerDoesntFinishQuickly).