Fixed python scripts to propagate 'hosts' var
In the case where users do not have access to the RAMCloud internal
rcres tool, our python scripts require the existence of a 'hosts'
variable in the user's localconfig.py to determine which machines to
use for a RAMCloud cluster. Unfortunately, this variable was not
properly propagated (imported) in some of our scripts which caused
erroneous warnings and failures. This is now fixed.