Fixed intermittent segfault in MembershipServiceTest.updateServerList_single.
The failure has only been reproduced on slower virtual machines. The problem is
that when the unit test function reaches its end, the CoordinatorService object
and the CoordinatorUpdateManager object inside happen to be destructed before
the CoordinatorServerList object. However, the updater thread in
CoordinatorServerList may still try to access the update manager in
CoordinatorServerList::pruneUpdates(), which results in the segfault.

The solution is to embed the CoordinatorServerList object inside the
CoordinatorService object and declare it after the update manager.