Partially replaced the index B+tree with SkipLists
A SkipList is a probabilistic indexing structure that consists of
multiple, hierarchical singly linked lists. The base level linked
list contains all the index entries and each successive level
contains fewer entries (i.e. 'skip'-ing certain entries). The main
advantage of this structure to other solutions like a B+tree is
simplicity; SkipLists do not require complicated balancing logic
and thus can be implemented in just a few hundred lines of code.

This commit contains a prototype of a SkipList implementation to
test the latency of lookup and writes. Unfortunately, we have found
that the SkipList scales rather poorly as the number of object
increases due to cache misses and lookups into ObjectManager
when traversing the links of the singly linked lists. In contrast,
the B+tree was found to be more cache-efficient as it batches 8-16
keys in a single Node read (the SkipList has one key per linked
list node). As a comparison point, the end-to-end latency of a
point lookup in an index containing 100k objects is about 11.5µs
for the B+tree vs 20.1µs for the SkipList.

Note that since this is just a prototype system for testing the
read/write latencies of SkipLists, not all of the indexing
features in RAMCloud have been implemented. Most notably, index
migration is broken and will cause segfaults when attempted.
Correspondingly, the following unit tests involving migration also
fail: MasterServiceTest.splitAndMigrateIndexlet*,
TableManagerTest.splitAndMigrateIndexlet, and
TableManagerTest.notifyReassignIndexlet.

One should only use this branch for experimental testing.