indexWriteDist changed to measure overwrite times
ClusterPerf's indexWriteDist used to fill a table with 1M objects
and then perform --count NEW writes ontop of the 1M objects. This
design deviates from writeDistRandom which performs overwrites on
the original 1M objects and more importantly, causes the Btree to
get slower and slower with increasing --count (due to more levels
being created). This patch changes indexWriteDist to fill the
table to 1M objects and then measure overwrites for the benchmark.