Persist ParticipantList in registerTransaction
The participant list can now be tracked in the TransactionManager.
Under the previous model, managing the persistence of the participant
list was done by ObjectManager which would have require a tight
coupling between the TransactionManager and ObjectManager with the
logic split between the two modules.  With this commit, the
registerTransaction method can now also manage the persistence of the
participant list which can be used in the following cases:
 - insert new participant list
 - recover existing participant list
 - migrate existing participant list
TransactionManager support for cleaning participant list entries is
not included.