Add InProgressTransaction::handleTimerEvent
The TransactionManager::InProgressTransaction::handleTimerEvent method
is called when a transaction has timed out because the transaction did
not complete in a timely manner.  The method will perform some
incremental progress toward initiating transaction recovery and will
reschedule itself to run in the future if more work needs to be done.