Reduce WorkerTimer::Manager::checkTimers overhead
WorkerTimer::Manager now keeps the timers in sorted order.  The sorted
insert causes WorkerTimer::start() and WorkerTimer::stop() to have
complexity logarithmic with the number of active timers, but reduces
the complexity of checkTimers down from linear to constant.

Previously, WorkerTimer::Manager kept timers in a queue which allowed
for fast WorkerTimer::start() calls but required a linear search to
find the next runnable timer.  This caused checkTimers to hold the
WorkerTimer::mutex for long periods of time causing an increase in the
latency of other WorkerTimer methods as they wait on the mutex.

With this commit, the longest any method may need to wait for the
mutex should be logarithmic with the number of active timer instead
of linear.