Removed unit tests from the Travis CI pipeline
The original goal of unit tests in Travis was to help detect when
developers have forgotten to run unit tests before committing, but
maintaining the unit tests in Travis with its intermittent and load
dependent failures has proved to be more effort than was worth for
the benefit for it was providing. Thus, this commit removes the unit
tests but keeps all the build tests within Travis.