Modified UdpDriver to receive packets in the background
In the recent refactoring of the Driver interface I moved the kernel
calls into the polling loop. This turned out to be a bad idea, because
when servers use basic+infud, they use basic+udp for communication with
the coordinator, so the slow kernel calls for UDP were damaging infud
performance also. This commit introduces a background thread to invoke
the recvmmsg kernel call.