IndexRpcWrapper API changed to require Context.
IndexRpcWrapper had two constructors, which required pointers to the MasterService
or the Ramcloud process invoking the constructor. It then used this information
to fetch the context. These constructors are now replaced by a single constructor
that directly requires the context.

This simplifies code, reduces redundancy and the number of forward declarations.