Refactor remove/write to correct locking behavior
ObjectManager holds the lock on the hash table bucket while removing
or writing and object. Move the code for updating index entries
(corresponding to the object being removed / written) such that it
executes while the lock is held for that object.

Add ObjectUpdate class that provides a callback which is used by
ObjectManager to notify MasterService that it is safe to respond
to an RPC. This is used in situations where the ObjectManager has
fairly expensive clean up to perform which needs the lock but for
which the RPC doesn't need to be delayed.
For example, while removing an object, ObjectManager invokes the
callback after it has removed the object from the master's log and
hash table, but before deleting the index entries corresponding to it.
Subclasses of ObjectUpdate overrride the callback to define
appropriate behavior for a given operation.

Change the setup of ObjectManagerTest to construct a mock cluster
and modify the unit tests accordingly.