Simplify TransactionManager Cleaner
The TransactionManager::TransactionRegistryCleaner is responsible for
removing unnecessary TransactionRecords from the TransactionManager.
Previously, the cleaner was invoked/scheduled on specific records when
necessary in an ad hoc fashion.  These ad hoc calls needed to be
carefully placed to ensure that no records would be "leaked;" this
required additional complex reasoning to ensure correctness.  Instead,
the cleaner now operates on the full collection of TransactionRecords
and is invoked periodically and regularly.  This make reasoning about
the cleaner's correctness simpler.