Handle non-existent table in transaction recovery
Previously, transactions would be left incomplete if a related table
were dropped during the processing of the transaction.  While we can't
ensure fully consistent transactions in the event of a table drop,
this patch ensures that the objects locked by the transaction will be
unlocked and the transaction's metadata will be cleaned up.

RAM-859 #close