Revert "Refactor remove/write to correct locking behavior"
This reverts commit 9fbf8ed331c3be4a83a4a5d82bba19f3b31ac16a.

Reverting this commit since this might degrade performance and it doesn't
really solve "safety" problem in face of network reordering. Details on problem
and possible solution are below.

1. Performance: I think holding bucket lock during index insertion and
deletion is not ideal. The entire bucket will be blocked for ~40us
(15 for insert, 10 for sync to backup, 15 for deletion). Even when we are not
using indexing, this commit forces to hold bucket lock during sync to backup.

2. Not Enough Safety: Holding bucket lock until index deletion actually
doesn't give safety in face of network packet re-ordering.
The first request of index delete can be held at some switch for long time, then
the master will send a retry of the same request. As the retry is processed,
the master will release the bucket lock. Then the original request may arrive at
index server, which will delete the index erronously.

Without implementing RIFL for indexing (probably too complex), I think the
easiest solution is keeping object version number in index entry or nonce in
both index entry and object. And we only remove index entry if the version or
nonce matches.
I thought reference counting can be easy solution but it also doesn't protect
us from the stale retry problem.

(Resolved) Conflicts:
	src/MasterService.cc
	src/ObjectManager.cc
	src/ObjectManagerTest.cc