Fix race conditions in WorkerTimerTest.
Currently, multiple test cases in WorkerTimerTest invoke
waitForWorkerProgress(), and then immediately attempt to read the TestLog.
Unfortunately, this read races with the thread managed by WorkerTimer::Manager,
which might have made progress but not yet outputted the TestLog.

More specifically, a return from waitForWorkerProgress is a necessary but not
sufficient condition for contents to appear in the TestLog.

This commit fixes this by adding a TestUtil::waitForLog(NULL); statement before
attempting to read the TestLog.