Add optional migration and stats to readDistWorkload.
This makes several small changes to readDistWorkload but it tries to
preserve the default way of doing things wherever possible.

The workload can now be limited by --seconds instead of a count. Samples
are taken continously, and the samples array is pre-sized upfront based
on --seconds and --targetOps.

Finally, there is an alterate sample output that includes the time at
which the sample was taken and not just the duration of the operation.
By default, the test outputs the old format and the old cdf/rcdf
extractions should work.  If --fullSamples is used the full set of
samples and an rcdf of the data are dropped into logs/latest.