Fix race between AbstractLog append() and free()
AbstractLog::append() and AbstractLog::free() both update the variable
AbstractLog::totalLiveBytes.  This variable is used to determine if
the log has space to append more data.  Previously, the append() and
free() calls raced to update totalLiveBytes.  This race condition causes
the variable to incorrectly track log usage and thus causes the log to incorrectly determine whether or not it can append more data.  This
commit fixes the race to update totalLiveBytes.

RAM-864 #close