Decoupled the TESTING flag from the DEBUG flag.
Before this commit, the TESTING flag is set to 1 if the DEBUG flag is equal to
`yes`. This setting has at least two problems: 1) we cannot run unit tests
under release build and 2) DpdkDriver is broken under debug build because its
mock implementation for unit tests is enabled by the TESTING flag.