common.py now returns a copy of the hosts list.
cluster.py removes an item from cluster.hosts when the disjunct option is passed.
If this is a reference to the user-defined `hosts` list instead of a copy, this
will result in subsequent call to checkHost() incorrectly believing that the
host is not valid because it has been removed from the list. This commit fixes
this problem.