Fixed gradle build scripts bug for java bindings.
Normally RAMCloud compiles its binaries and other files to a folder
named "obj.<branch-name>", but when in detached head state it compiles
these files to the folder "obj". The gradle build scripts weren't
handling this case. This change detects this state, and directs the
gradle build scripts to find necessary files in "obj".