Fixed an unsigned integer underflow bug in QueueEstimator
... that results in the queue estimator constantly reporting the transmit
queue being empty even though it's not. The underflow bug could be triggered
when getQueueSize(time) is invoked with a stale timestamp smaller than the
stored `currentTime`. When `time` is substracted by `currentTime`, because of
the underflow, we get a large unsigned interger as the time passed and then
a negative number as the new queue size, suggesting the tx queue is empty.