Integrated spdlog into RAMCloud
It's possible to supplant RAMCloud's own logger with spdlog by
compiling with "SPDLOG=yes". Note that this mode completely
replaces the RAMCloud logger for all logging except clogs,
backtraces, or die's and is INCOMPATIBLE with "NANOLOG=yes"
(though regular NANO_LOG invocations can be used). The spdlog
log files are appended with a ".spdlog" extension.