Changed DPDK driver to wait for link to come up.
Previously the DPDK driver would call rte_eth_dev_get_nowait immediately after
calling rte_eth_dev_start, and exit with an exception if it found that the link
status was not UP. For the Mellanox Connect-X3 NICs this seemed to work fine,
but the Intel X710 NICs only change their link status after a short delay.

Changed this call from rte_eth_dev_get_nowait to rte_eth_dev_get, which will
wait for up to 9 seconds for the link status to change to UP before returning.
If the link status is still not up after 9 seconds, then the driver will exit
with an exception just as before.