Moved MTU config to happen before device start.
Previously, the DPDK driver was setting the MTU on the device *after*
the device was started. This worked for Mellanox Connect-X3 NICs, but
returned with an error for Intel X710 NICs, complaining that the MTU can
not be set after the device is started.

Moved MTU to be set before the NIC device is started.