Change multiple independent tx queues to a single tx queue that communicates with loopback, but that does not seem to be the problem.
It seems to make read latency higher since there's now more work for the rx
thread, but it doesn't help writeLatency. I wonder if we are simply shoving
more bytes at the NIC than it can handle, and this is causing immense delays
because the NIC has to buffer it in some very slow memory.