Fixed the client-side response buffer corruption bug in BasicTransport.
Problem: I broke the code in MessageAccumulator::addPacket accidentally
by using appendExternal to append packet buffers to the message buffer.
This is bad since it doesn't transfer the memory ownership to the buffer.
Therefore, when the Transport's ClientRpc oject (owner of the accumulator)
is destroyed, the ownership of the packet buffers are returned to the
driver prematurely, while the client's RPC object is still alive.

Fix: Use Driver::PayloadChunk::appendToBuffer to append packet buffers
and implement the incremental release of packet buffers at the driver
layer (as opposed to the transport layer).