A few TimeTrace enhancements
The performance of TimeTrace::record in practice could be much worse than
the number given by Perf.cc due to less effective pipelining. For example,
I have observed that the average gap between consecutive TimeTrace::record
statements in HomaTransport to be ~15 ns. With the optimizations in this
commit, the average gap is reduced to ~10 ns.

Changes:
* Print thread ID at the beginning of each timetrace entry
* Move TimeTrace::Buffer::record implementation to the header to allow inlining
* Use variadic templates in TimeTrace::record to avoid unnecessary assignments
when the number of event args is less than 4.