Fixed TxHintFailedRpc construction bug.
TxHintFailedRpc's participant list payload is appended to its request buffer using Buffer.append()
in the constructor, which was doing a copy-free append. However, the caller
(TransactionManager::TransactionRecord::handleTimerEvent()) was handing a participant list pointer
to the constructor that was ultimately owned by a stack resident Buffer. Therefore, if
TransactionManager::TransactionRecord::handleTimerEvent() returned before the RPC was sent, then
the participant list payload was at risk of being corrupted, which we observed in practice.

Fixed this by changing the Buffer.append() call to a Buffer.appendCopy() call.